Fix exit button in Library
When library screen starts, keyboard_event_queue is called to enable
SWITCH_1 and SWITCH_5. When SWITCH_5 is pressed, screen_library_content
disables SWITCH_1. For content with more than one page, the switch is
re-enabled by Display_library_content(), but for content with a single
page, it returns to screen_home_entry with SWITCH_1 still disabled,
which causes the exit function to stop working.

This PR resets the state of the enabled buttons upon returning from the
content screen.

Signed-off-by: Rodrigo Chiossi <rodrigochiossi@gmail.com>